+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.2.0 (Feb 18)
+---
+- Add more configuration options namely, ```model```, ```useSpan```, ```imageList``` and ```expanded```
+- TreeNode icons
+
 0.1.5 (Feb 13)
 ---
 - Load tree from data Model